opt(all): 使用double保存GPS秒信息.

+++ bsp_config.h
@@ -26,7 +26,7 @@ public:
     struct Gps_Info
     {
         qint32 week;
-        qint32 current_week_ms;
+        double current_week_ms;
         double latitude;
         double longitude;
         double altitude;
